pyproject.toml
[project]
name = "create-olyv-app"
version = "0.1.0"
description = "Quickly scaffold a new Django project with Olyv's opinionated setup and best practices."
readme = "README.md"
license = { file = "LICENSE" }
authors = [
    { name = "Christian Who Codes", email = "developer@christianwhocodes.space" }
]
requires-python = ">=3.13"
dependencies = [
    "termcolor>=3.1.0",
]

[build-system]
requires = ["uv_build>=0.8.18,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "olyv",
    "djlint>=1.36.4",
    "python-decouple>=3.8",
]

[tool.uv.sources]
olyv = { path = "../olyv", editable = true }

[project.scripts]
create-olyv-app = "create_olyv_app.cli:main"
